+++ recursion/level01/get/MazePathWithJumps.java
@@ -1,4 +1,4 @@
-package recursion.level01;
+package recursion.level01.get;
 
 import java.util.ArrayList;
